.requirements.txt
pytest
pyyaml
jsonschema
ga4gh.gks.metaschema
